Conflicts:
	guestgreen/src/main/java/kr/co/green/main/controller/MainController.java
	guestgreen/src/main/java/kr/co/green/main/model/dao/MainDAO.java
	guestgreen/src/main/java/kr/co/green/main/model/service/MainService.java
	guestgreen/src/main/java/kr/co/green/main/model/service/MainServiceImpl.java
	guestgreen/src/main/java/kr/co/green/project/controller/ProjectEnrollController.java
	guestgreen/src/main/java/kr/co/green/project/model/dao/ProjectDAO.java
	guestgreen/src/main/java/kr/co/green/project/model/dto/ProjectDTO.java

+++ guestgreen/src/main/java/kr/co/green/project/model/dto/ProjectDTO.java
@@ -10,7 +10,7 @@ public class ProjectDTO {
 	private int projectPrice;
 	private int projectTargetAmount;
 	private int projectCurrentAmount;
-	private int projectCurrentPercentage;
+	private double projectCurrentPercentage;
 	private int projectSponserNumber;
 	private String projectConfirmStatus;
 	private String projectSaleStatus;
@@ -244,11 +244,12 @@ public class ProjectDTO {
 		this.projectManagerImagePath = projectManagerImagePath;
 	}
 
-	public int getProjectCurrentPercentage() {
+	public double getProjectCurrentPercentage() {
+
 		return projectCurrentPercentage;
 	}
 
-	public void setProjectCurrentPercentage(int projectCurrentPercentage) {
+	public void setProjectCurrentPercentage(double projectCurrentPercentage) {
 		this.projectCurrentPercentage = projectCurrentPercentage;
 	}
